preparing LinkActors and DoughnutActors
