+++ LuckParser/LuckParser.csproj
@@ -120,6 +120,7 @@
     <Compile Include="Models\DataModels\ParsedLog.cs" />
     <Compile Include="Models\DataModels\SettingsContainer.cs" />
     <Compile Include="Models\DataModels\Statistics.cs" />
+    <Compile Include="Models\ParseModels\CombatReplay\Actors\Actor.cs" />
     <Compile Include="Models\ParseModels\CombatReplay\Actors\CircleActor.cs" />
     <Compile Include="Models\ParseModels\CombatReplay\Actors\FollowingCircle.cs" />
     <Compile Include="Models\ParseModels\CombatReplay\Actors\ImmobileCircle.cs" />
